Renamed log to utilities.py, moved csv to dataframe function there to avoid redundancy
