+++ pipeline/extract.py
@@ -5,7 +5,7 @@ import csv
 from dotenv import load_dotenv
 import requests
 
-from log import get_logger, set_logger
+from utilities import get_logger, set_logger
 
 
 class PlantAPIClient:
