+++ pipeline/load.py
@@ -1,23 +1,10 @@
 """Load script for inserting minute-by-minute data generated by transform into database."""
 
 from os import environ as ENV
-import pandas as pd
 from dotenv import load_dotenv
 import sqlalchemy
 
-from log import get_logger, set_logger
-
-
-def load_data(file_path: str = 'data/normalised_output.csv') -> pd.DataFrame:
-    """Loads the unclean data from csv."""
-    logger = get_logger()
-    if not isinstance(file_path, str):
-        logger.critical("Invalid URL type.")
-        raise TypeError("Invalid URL type.")
-    if file_path[-4:] != ".csv":
-        logger.critical("Filename doesn't end in .csv.")
-        raise ValueError("Please end your filename in .csv.")
-    return pd.read_csv(file_path)
+from utilities import get_logger, set_logger, load_csv_data
 
 
 def insert_transformed_data() -> None:
@@ -26,7 +13,7 @@ def insert_transformed_data() -> None:
     logger = get_logger()
 
     logger.info("Loading local recent data...")
-    transformed_data = load_data()
+    transformed_data = load_csv_data()
     logger.info("Successfully loaded recent data!")
 
     logger.info("Inserting data into database setup...")
