+++ pipeline/transform.py
@@ -2,19 +2,7 @@
 
 import pandas as pd
 
-from log import get_logger, set_logger
-
-
-def load_data(file_path: str = 'data/output.csv') -> pd.DataFrame:
-    """Loads the unclean data from csv."""
-    logger = get_logger()
-    if not isinstance(file_path, str):
-        logger.critical("Invalid URL type.")
-        raise TypeError("Invalid URL type.")
-    if file_path[-4:] != ".csv":
-        logger.critical("Filename doesn't end in .csv.")
-        raise ValueError("Please end your filename in .csv.")
-    return pd.read_csv(file_path)
+from utilities import get_logger, set_logger, load_csv_data
 
 
 def clean_dataframe(file_path: str = 'data/output.csv') -> pd.DataFrame:
@@ -23,7 +11,7 @@ def clean_dataframe(file_path: str = 'data/output.csv') -> pd.DataFrame:
     logger = get_logger()
     logger.info("Loading data from file path...")
     try:
-        input_dataframe = load_data(file_path)
+        input_dataframe = load_csv_data(file_path)
     except Exception as exc:
         logger.info("Data did not successfully load.")
         raise exc
